Document what we do in terms of symlinks and project root paths
It's not the job of this plugin to attempt to "follow" symlinks by
closing/reopen buffers to get to the resolved filename, just to work out
what project it is part of. Perhaps we will make this configurable
later, but not now.

So if you have a file file-a.txt in ~/ProjectA and you add a symlink to
it in ~/ProjectB then when you edit the symlink it will appear to be
part of ~/ProjectB instead.